Small code refactor changes Added PIPE Library (thanks to @natefinch 's library https://github.com/natefinch/npipe) Added SDDL Parser and Linux chmod parser for pipes/sockets Added Guardian/Sentinel structs to keep processes alive via Pipes Added AdjustPrivileges for Windows, which takes privilege names Fixed 32bit compatibility issues with the data package Added a task for running shellcode Added the 'Elevated' parent flag, which will choose only processes with a elevated token (windows only) Updated the XOR alias to fix typecasting for the cipher.Block interface Updated an error in crypto.CBK where the BlockSize() function would return -1 Small code optimizations other stuff idk lol

+++ c2/config.go
@@ -519,10 +519,10 @@ func ConnectWC2(url, agent, host string) Setting {
 	if len(h) > 255 {
 		h = h[:255]
 	}
-	if len(a) > data.DataLimitMedium {
+	if uint64(len(a)) > data.DataLimitMedium {
 		a = a[:data.DataLimitMedium]
 	}
-	if len(u) > data.DataLimitMedium {
+	if uint64(len(u)) > data.DataLimitMedium {
 		u = u[:data.DataLimitMedium]
 	}
 	s := Setting{wc2ID, byte(len(a) >> 8), byte(len(a)), byte(len(u) >> 8), byte(len(u)), byte(len(h))}
